# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ROSA Regional Frontend API
  description: Frontend API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /authz/policies:
    post:
      summary: Create a policy
      description: |
        Creates a new policy template. Policies define permissions using an IAM-like
        format with statements containing effects, actions, resources, and conditions.
        Policies must be attached to users or groups to take effect.
      operationId: createPolicy
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
      responses:
        '201':
          description: Policy created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          description: Bad request - invalid policy format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - account not provisioned or caller not admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List policies
      description: Returns a list of all policies for the account.
      operationId: listPolicies
      tags:
        - Authorization
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyList'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /authz/policies/{id}:
    get:
      summary: Get a policy
      description: Returns a specific policy by ID.
      operationId: getPolicy
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Policy ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Policy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a policy
      description: |
        Updates an existing policy. All attachments using this policy will be
        updated with the new policy content.
      operationId: updatePolicy
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Policy ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePolicyRequest'
      responses:
        '200':
          description: Policy updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          description: Bad request - invalid policy format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a policy
      description: |
        Deletes a policy. All attachments using this policy will also be deleted.
      operationId: deletePolicy
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Policy ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Policy deleted successfully
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Authorization - Group Management
  /authz/groups:
    post:
      summary: Create a group
      description: Creates a new group that can contain multiple users.
      operationId: createGroup
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Group with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List groups
      description: Returns a list of all groups for the account.
      operationId: listGroups
      tags:
        - Authorization
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupList'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /authz/groups/{id}:
    get:
      summary: Get a group
      description: Returns a specific group by ID.
      operationId: getGroup
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a group
      description: |
        Deletes a group and all its member associations.
        Attachments to this group will also be deleted.
      operationId: deleteGroup
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Group deleted successfully
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /authz/attachments/{id}:
    delete:
      summary: Delete an attachment
      description: |
        Detaches a policy from a user or group. The Cedar policy in
        Amazon Verified Permissions will be deleted.
      operationId: deleteAttachment
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Attachment ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Attachment deleted successfully
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Authorization - Admin Management
...
  /authz/admins/{arn}:
    delete:
      summary: Remove an admin
      description: Removes a principal from the admin list.
      operationId: removeAdmin
      tags:
        - Authorization
      parameters:
        - name: arn
          in: path
          required: true
          description: Admin's ARN (URL-encoded)
          schema:
            type: string
      responses:
        '204':
          description: Admin removed successfully
        '403':
          description: Forbidden - caller is not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      required:
        - code
        - reason
      properties:
        id:
          type: string
          description: Error ID
        kind:
          type: string
          description: Resource kind
          example: Error
        href:
          type: string
          description: API href reference
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
        operation_id:
          type: string
          description: Request operation ID for tracing
...
    CreatePolicyRequest:
      type: object
      description: Request body for creating a policy
      required:
        - name
        - policy
      properties:
        name:
          type: string
          description: Policy name
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Optional policy description
          maxLength: 1024
        policy:
          $ref: '#/components/schemas/V0Policy'
...
    Policy:
      type: object
      description: A policy template
      required:
        - id
        - kind
        - href
        - name
        - policy
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Policy ID
        kind:
          type: string
          description: Resource kind
          example: Policy
        href:
          type: string
          description: API href reference
        name:
          type: string
          description: Policy name
        description:
          type: string
          description: Policy description
        policy:
          $ref: '#/components/schemas/V0Policy'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    PolicyList:
      type: object
      description: Paginated list of policies
      required:
        - kind
        - page
        - size
        - total
        - items
      properties:
        kind:
          type: string
          example: PolicyList
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
...